+++ client/test/node/reader/DecisionReviewer-test.js
@@ -256,6 +256,42 @@ describe('DecisionReviewer', () => {
   });
 
   context('PDF list view', () => {
+    context('last read indicator', () => {
+      it('appears on latest read document', asyncTest(async() => {
+        // Click on first document link
+        wrapper.find('a').findWhere(
+          (link) => link.text() === documents[0].type).
+          simulate('mouseUp');
+        await pause();
+
+        // Next button moves us to the next page
+        wrapper.find('#button-next').simulate('click');
+        await pause();
+
+        wrapper.find('#button-backToDocuments').simulate('click');
+        // Make sure that the 1st row has the last
+        // read indicator in the first column.
+        expect(wrapper.find('#table-row-1').childAt(0).
+          children()).to.have.length(1);
+      }));
+
+      it('appears on document opened in new tab', asyncTest(async() => {
+        const event = {
+          ctrlKey: true
+        };
+
+        wrapper.find('a').findWhere(
+          (link) => link.text() === documents[0].type).
+          simulate('mouseUp', event);
+        await pause();
+
+        // Make sure that the 0th row has the last
+        // read indicator in the first column.
+        expect(wrapper.find('#table-row-0').childAt(0).
+          children()).to.have.length(1);
+      }));
+    });
+
     context('when expanded comments', () => {
       it('can view comments', () => {
         expect(wrapper.text()).to.not.include('Test Comment');
